clients/soa/python/requirements.txt
-e ./
-e ../../api/python
pytest
pytest-asyncio
pytest-aiohttp
tox